+++ NEWS.md
@@ -1,3 +1,19 @@
+# CHANGES IN litedown VERSION 0.4
+
+- Provided an option `options(litedown.roam.cleanup = TRUE)` to clean up the `*__files/` directory after previewing `.Rmd` or `.R` files via `litedown::roam()` (thanks, @TimTaylor, #36).
+
+- Added the keyboard shortcut `Ctrl + K` (or `Command + K` on macOS) for rendering a file to disk in the `litedown::roam()` preview.
+
+- Cross-references also work for LaTeX output now.
+
+- Fixed an error in the internal function `detect_pkg()` during `R CMD check` on CRAN.
+
+- Set `options(bitmapType = 'cairo')` on macOS only when `xquartz` is available. Previously only `capabilities('cairo')` was checked, which was not enough. This option can also be manually set via `options(bitmapType)` in a code chunk if the automatic switch to `cairo` is not desired.
+
+- Fixed the bug that indented or quoted code blocks are not correctly indented or quoted when a code expression contains multiple lines.
+
+- Fixed the bug that the span syntax `[text](){...}` doesn't work when `text` contains markup (e.g., bold or italic).
+
 # CHANGES IN litedown VERSION 0.3
 
 - Added a new engine `md` to output Markdown text both verbatim and as-is, which can be useful for showing Markdown examples, e.g.,